First step in simplifying webpack pipeline.
